Fix Bot MediaStream initialization
The BotMediaStream, wrapping the media sockets, needs to be created first. This will ensure that all the media events are subscribed to before the call gets processed.
Avoiding a race condition, where the sends status events on the sockets are raised after the CallsOnUpdated with added notification is raised.
Media can start flowing before an Establishing notification is sent to the bot.